+++ src/re_db/d.clj
@@ -11,9 +11,6 @@
        (do ~@body)
        @~'re-db.core/*db-log*)))
 
-(defmacro capture-patterns [& body]
-  `(~'re-db.core/capture-patterns ~@body))
-
 (defmacro try-branch [db on-error & body]
   `(let [db-before# ~db]
      (try (branch db-before# ~@body)