+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 ## Changelog
 
 ### 0.1.9
-- All DB listeners are now called with the tx-report map, instead of data related to the pattern(s) the listener was registered on.
-- `listen!` takes a vector of patterns instead of multiple args.
+- Listeners are called with the tx-report map instead of data related to the pattern(s) the listener was registered on.
+- `listen!` now accepts a map of patterns, ie. {<kind>, <collection of patterns>}, instead of multiple patterns as args, for clarity and performance.
+   To register a :tx-log listener, don't the 2-arity version.
+   Pattern maps returned from the `capture-patterns` macro can be passed directly to `listen!`.
+- Pattern-related functions moved to re-db.patterns namespace
 - `entity-ids` and `entities` take a vector of patterns instead of multiple args.
 
 ### 0.1.8
 
-- The `:eav` index is now a sorted-map (from `clojure.data.avl`) which supports `subrange`. This is primarily to support path-based id attributes for hierarchical schemas.
 - Overall transaction time reduced by ~60%, w/ a further 30% reduction with option `{:notify false}`, which supercedes the :mute option.
 - :cardinality/many attributes must be passed as sets, eg/
   [:db/add 1 :my-attr #{"new-val"}] or [:db/retract-attr 1 :my-attr #{"old-val"}].